04- Deep Learning/Titanic_ANN_Project/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
scikit-learn==1.7.2
python-dotenv==1.2.1
python-multipart==0.0.20
matplotlib==3.10.7
seaborn==0.13.2
tensorflow==2.20.0
joblib==1.5.2
pydantic==2.12.5
